pyproject.toml
[project]
name = "api-exercise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "mysql-connector-python>=9.3.0",
    "python-dotenv>=1.1.0",
]
